Excluded all that is not needed from CLMNCEP case
This is done in datm_comp_run() when COUP_OAS_ICON is set.  Model concludes
without errors when all streams are included.  However, before datm_comp_run()
and before oas_receive_icon(), lnd_import() is executed.  That routine includes
checks like long-wave radiation from the atmosphere not being zero or negative:
it fails on this when the redundant streams are not read.