Excluded all that is not needed from CLMNCEP case
This is done in datm_comp_run() when COUP_OAS_ICON is set.  Model concludes
without errors when all streams are included.  Also when Precip or Solar are
excluded, it concludes without errors, but excluding TPQW results in a crash
close to the end of a 23-model-h simulation (see issue 74).